requirements.txt
streamlit
numpy
opencv-python
Pillow
tensorflow
matplotlib
google-cloud-aiplatform #this is required
google-cloud-logging  #this is required